+++ projects/extension/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- Add "latest block" information to Popup, ([#1254](https://github.com/paritytech/substrate-connect/pull/1254))
+
 ## 0.1.10 - 2022-09-15
 
 ### Changed